splitting out database files
